Revert "Revert "Revert "plusieur changement a merge"""
This reverts commit 1823bf6e4c89d83e86f27b10b46fcbc204fe936f.

+++ lescitronsdetoto-backend/queries/InterventionQueries.js
@@ -1,4 +1,3 @@
-
 const HttpError = require('../HttpError');
 const pool = require('./DBPool');
 
@@ -6,7 +5,7 @@ const getAllInterventions = async () => {
     const result = await pool.query(
         `SELECT id_intervention, type_intervention, valeur_intervention, etat_intervention
          FROM intervention
-         ORDER BY id_intervention`
+         ORDER BY id_intervention`,
     );
 
     return result.rows.map((row) => {
@@ -14,7 +13,7 @@ const getAllInterventions = async () => {
             idIntervention: row.id_intervention,
             typeIntervention: row.type_intervention,
             valeurIntervention: row.valeur_intervention,
-            etatIntervention: row.etat_intervention,
+            etatIntervention: row.etat_intervention
         };
 
         return intervention;
@@ -22,63 +21,91 @@ const getAllInterventions = async () => {
 };
 exports.getAllInterventions = getAllInterventions;
 
-const getIntervention = async (idIntervention) => {
-    const client = await pool.connect();
+const getIntervention = async (idIntervention, clientParam) => {
+    const client = clientParam || (await pool.connect());
 
     try {
-        await client.query('BEGIN');
+        if (!clientParam) {
+
+            await client.query("BEGIN");
+        }
 
         const result = await client.query(
             `SELECT id_intervention, type_intervention, valeur_intervention, etat_intervention
-            FROM intervention
-            WHERE id_intervention = $1`,
+         FROM intervention
+         WHERE id_intervention = $1`,
             [idIntervention]
         );
 
         const row = result.rows[0];
         if (row) {
-            return {
+            const intervention = {
                 idIntervention: row.id_intervention,
-                typeIntervention: row.type_intervention,
-                valeurIntervention: row.valeur_intervention,
-                etatIntervention: row.etat_intervention,
-            };
-        }
-        return undefined;
+            typeIntervention: row.type_intervention,
+            valeurIntervention: row.valeur_intervention,
+            etatIntervention: row.etat_intervention
+            }
+
+
+
+            return intervention;
+        };
     } catch (err) {
-        await client.query('ROLLBACK');
+        if (!clientParam) {
+            await client.query("ROLLBACK");
+        }
         throw err;
     } finally {
-        client.release();
+        if (!clientParam) {
+            client.release();
+        }
     }
 };
 exports.getIntervention = getIntervention;
 
-const createIntervention = async (intervention) => {
-    const client = await pool.connect();
+
+
+const createIntervention = async (intervention, clientParam) => {
+    const client = clientParam || (await pool.connect());
 
     try {
-        await client.query('BEGIN');
+        if (!client) {
+            await client.query('BEGIN');
+        }
+
 
         const result = await client.query(
-            `INSERT INTO intervention (type_intervention, valeur_intervention, etat_intervention)
-            VALUES ($1, $2::money, $3)
-            RETURNING id_intervention`,
-            [intervention.typeIntervention, intervention.valeurIntervention, false]
+            `INSERT INTO intervention ( type_intervention, valeur_intervention, etat_intervention ) 
+                         VALUES ($1, $2, $3 )
+                         RETURNING id_intervention`,
+            [
+                
+                intervention.typeIntervention,
+                intervention.valeurIntervention,
+                intervention.etatIntervention,
+            
+            ]
         );
 
-        if (result.rows.length === 0) {
-            throw new HttpError("Impossible de créer l'intervention", 500);
+        const newIntervention = await getIntervention(result.rows[0].id_intervention, client);
+        if (!client) {
+            await client.query('COMMIT');
         }
-        return result.rows[0].id_intervention;
+
+        return newIntervention;
     } catch (err) {
-        await client.query('ROLLBACK');
-        throw new HttpError("Une erreur est survenue lors de la création de la transaction ", 500);
+        if (!client) {
+            await client.query('ROLLBACK');
+        }
+        throw new HttpError("Une erreur est survenue lors de la création de l'employé", 500);
     } finally {
-        client.release();
+        if (!client) {
+            client.release();
+        }
     }
 };
 exports.createIntervention = createIntervention;
+
 
 const updateIntervention = async (intervention) => {
     const client = await pool.connect();
@@ -87,25 +114,26 @@ const updateIntervention = async (intervention) => {
         await client.query('BEGIN');
 
         const result = await client.query(
-            `UPDATE intervention
-            SET type_intervention = $2, valeur_intervention = $3, etat_intervention = $4
+            `UPDATE intervention SET  type_intervention = $2, valeur_intervention = $3, etat_intervention = $4 
             WHERE id_intervention = $1`,
             [intervention.idIntervention, intervention.typeIntervention, intervention.valeurIntervention, intervention.etatIntervention]
         );
         if (result.rowCount === 0) {
-            throw new HttpError(`Impossible de trouver l'intervention avec id_intervention ${intervention.idIntervention}`, 404);
+            throw new HttpError(`Impossible de trouver l'employé avec id_intervention ${intervention.idIntervention}`, 404);
         }
 
-        await client.query('COMMIT');
+        await client.query("COMMIT");
         return intervention;
     } catch (err) {
-        await client.query('ROLLBACK');
-        throw new HttpError(`Une erreur est survenue lors de la mise à jour de l'intervention avec id_intervention ${intervention.idIntervention}`, 500);
+        await client.query("ROLLBACK");
+        throw new HttpError(`Une erreur est survenue lors de la mise à jour de l'employé avec id_intervention ${intervention.idIntervention}`, 500);
     } finally {
         client.release();
     }
 };
 exports.updateIntervention = updateIntervention;
+
+
 
 const deleteIntervention = async (idIntervention, clientParam) => {
     const client = clientParam || (await pool.connect());
@@ -114,7 +142,34 @@ const deleteIntervention = async (idIntervention, clientParam) => {
         if (!clientParam) {
             await client.query('BEGIN');
         }
-        
+        // Vérifier d'abord si l'employé est lié à message_chat ou evenement
+        const checkQuery = `
+            SELECT id_intervention 
+            FROM intervention
+            WHERE id_intervention NOT IN (
+                SELECT id_intervention FROM vehicule_intervention
+
+            )
+        `;
+
+        const checkResult = await pool.query(checkQuery);
+
+        if (checkResult.rows.find(row => row.id_intervention === idIntervention)) {
+
+
+            // Si l'employé est lié, retourner une réponse vide
+            throw new error("L'employé est lier a une autre table");
+
+        }
+
+        // Supprimer de user_account
+        const deleteQuery = `
+            DELETE FROM user_account
+            WHERE id_intervention = $1
+        `;
+
+        await pool.query(deleteQuery, [idIntervention]);
+
         // Supprimer de la table intervention
         const deleteInterventionQuery = `
             DELETE FROM intervention
@@ -128,6 +183,8 @@ const deleteIntervention = async (idIntervention, clientParam) => {
         }
 
         return {};
+
+
     } catch (err) {
         await client.query('ROLLBACK');
         throw err;
@@ -139,3 +196,4 @@ const deleteIntervention = async (idIntervention, clientParam) => {
 };
 
 exports.deleteIntervention = deleteIntervention;
+
